moved task_data_factory and task_event_data_factory fixtures to app/test_task.py
