+++ tests/app/test_task.py
@@ -4,7 +4,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from corva.app.task import TaskApp
-from corva.models.task import TaskStatus
+from corva.models.task import TaskStatus, TaskData, TaskEventData
 from corva.models.task import UpdateTaskData
 from tests.conftest import ComparableException, APP_KEY, CACHE_URL
 
@@ -12,6 +12,40 @@ from tests.conftest import ComparableException, APP_KEY, CACHE_URL
 @pytest.fixture(scope='function')
 def task_app():
     return TaskApp(app_key=APP_KEY, cache_url=CACHE_URL)
+
+
+@pytest.fixture(scope='session')
+def task_data_factory():
+    def _task_data_factory(**kwargs):
+        for key, val in dict(
+             id=str(),
+             state='running',
+             asset_id=int(),
+             company_id=int(),
+             app_id=int(),
+             document_bucket=str(),
+             properties={},
+             payload={},
+        ).items():
+            kwargs.setdefault(key, val)
+
+        return TaskData(**kwargs)
+
+    return _task_data_factory
+
+
+@pytest.fixture(scope='session')
+def task_event_data_factory():
+    def _task_event_data_factory(**kwargs):
+        for key, val in dict(
+             task_id=str(),
+             version=2
+        ).items():
+            kwargs.setdefault(key, val)
+
+        return TaskEventData(**kwargs)
+
+    return _task_event_data_factory
 
 
 def test_group_by_field():
