+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 from fakeredis import FakeRedis
 
 from corva.event import Event
-from corva.models.task import TaskData, TaskContext, TaskEventData
+from corva.models.task import TaskContext
 from corva.state.redis_adapter import RedisAdapter
 from corva.state.redis_state import RedisState
 
@@ -46,40 +46,6 @@ def redis(redis_adapter):
 class ComparableException(Exception):
     def __eq__(self, other):
         return type(self) is type(other) and self.args == other.args
-
-
-@pytest.fixture(scope='session')
-def task_event_data_factory():
-    def _task_event_data_factory(**kwargs):
-        for key, val in dict(
-             task_id=str(),
-             version=2
-        ).items():
-            kwargs.setdefault(key, val)
-
-        return TaskEventData(**kwargs)
-
-    return _task_event_data_factory
-
-
-@pytest.fixture(scope='session')
-def task_data_factory():
-    def _task_data_factory(**kwargs):
-        for key, val in dict(
-             id=str(),
-             state='running',
-             asset_id=int(),
-             company_id=int(),
-             app_id=int(),
-             document_bucket=str(),
-             properties={},
-             payload={},
-        ).items():
-            kwargs.setdefault(key, val)
-
-        return TaskData(**kwargs)
-
-    return _task_data_factory
 
 
 @pytest.fixture(scope='session')
